aula4

+++ src/main/java/br/edu/iffarroupilha/autocentersb/model/ServicoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ServicoRepository extends JpaRepository<Servico, Integer> {
+public interface ServicoRepository extends JpaRepository<Servico, Long> {
     List<Servico> findByDescricao(String descricao);
 }
