+++ src/main/java/br/edu/iffarroupilha/autocentersb/model/service/ServicoService.java
@@ -22,7 +22,7 @@ public class ServicoService {
         return this.servicoRepository.findAll();
     }
 
-    public void delete(Integer idServico){
+    public void delete(Long idServico){
         this.servicoRepository.deleteById(idServico);
     }
